Split raylib-based camera/picking/misc-widget code into rl_utils.cpp/.h
Moves the ~1300 lines of camera/picking/mini-compass/misc-ImGui-widget
code (this app's raylib-based replacement for the shared core/src/utils.cpp
API) out of multi_view_tls_registration_gui.cpp into their own
rl_utils.cpp/rl_utils.h files, mirroring how core/src/utils.cpp /
Core/utils.hpp were structured for the GLUT apps.
